requirements.txt
amqp==5.2.0
asgiref==3.7.2
attrs==23.1.0
autobahn
Automat==22.10.0
billiard==4.2.0
celery==5.3.6
cffi==1.16.0
channels==4.0.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cryptography==41.0.5
daphne==4.0.0
Django==4.2.6
django-cors-headers==4.2.0
django-filter==23.3
django-js-asset==2.1.0
django-model-utils==4.3.1
django-mptt
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
gunicorn==21.2.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
kombu==5.3.4
packaging==23.2
Pillow==10.1.0
prompt-toolkit==3.0.41
psycopg2-binary==2.9.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==23.3.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
Twisted==23.10.0
txaio==23.1.1
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.12
zope.interface==6.1
google-auth==2.22.0
google-oauth2-tool==0.0.3
google==3.0.0